refactor: move sjcl.js back into /boot
