--- boot/lib/tiddlywiki.files
@@ -1,14 +1,0 @@
-{
-	"tiddlers": [
-		{
-			"file": "sjcl.js",
-			"fields": {
-				"title": "$:/library/sjcl.js",
-				"type": "application/javascript",
-				"library": "yes"
-			},
-			"prefix": "(function(define) {\n",
-			"suffix": "\n})(function (_,defined){window.sjcl = defined()})\n"
-		}
-	]
-}
